@@ -1,6 +1,5 @@
 const multer = require('multer');
 
-// TODO: Manually confirm files are saved to fs
 // TODO: Mock multer and write unit tests to confirm returned object's shape
 // Utilizes the 'multer' library to save files to disk (database storage not as performant).
 const uploadFile = () => {
@@ -11,11 +10,11 @@ const uploadFile = () => {
         },
         // Specify file name (appends current date)
         filename: function (req, file, cb) {
-            cb(null, file.fieldname + '-' + Date.now());
+            cb(null, Date.now() + '-' + file.originalname);
         }
     });
 
-    return multer({ storage: storage });
+    return multer({ storage });
 };
 
 module.exports = {
